refactor(identity-server-application): Token 발급 책임 정리
