--- identity/identity-server/identity-server-application/src/main/java/com/seatliberator/seatliberator/identity/server/application/jwks/port/in/TokenProvider.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.identity.server.application.jwks.port.in;
-
-import java.util.Map;
-
-public interface TokenProvider {
-    String issue(Map<String, Object> attributes);
-}
