--- identity/identity-server/identity-server-application/src/main/java/com/seatliberator/seatliberator/identity/server/application/jwks/service/OpaqueTokenProvider.java
@@ -1,21 +1,0 @@
-package com.seatliberator.seatliberator.identity.server.application.jwks.service;
-
-import com.seatliberator.seatliberator.identity.server.application.jwks.port.in.TokenProvider;
-import com.seatliberator.seatliberator.identity.server.application.jwks.port.out.ByteEncoder;
-import lombok.RequiredArgsConstructor;
-
-import java.util.Map;
-import java.util.random.RandomGenerator;
-
-@RequiredArgsConstructor
-public class OpaqueTokenProvider implements TokenProvider {
-    private final RandomGenerator randomGenerator;
-    private final ByteEncoder byteEncoder;
-
-    @Override
-    public String issue(Map<String, Object> attributes) {
-        byte[] randomBytes = new byte[32];
-        randomGenerator.nextBytes(randomBytes);
-        return byteEncoder.encode(randomBytes);
-    }
-}
