+++ identity/identity-server/identity-server-application/src/main/java/com/seatliberator/seatliberator/identity/server/application/token/internal/OpaqueTokenGenerator.java
@@ -1,6 +1,5 @@
 package com.seatliberator.seatliberator.identity.server.application.token.internal;
 
-import com.seatliberator.seatliberator.identity.server.application.jwks.port.out.ByteEncoder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
